Added tournament functionality.

+++ database/migrations/2024_01_15_000001_create_rankings2024s1_table.php
@@ -19,7 +19,7 @@ class CreateRankings2024s1Table extends Migration
             $table->string('network_group');
             $table->string('game');
             $table->integer('fase');
-            $table->integer('points');
+            $table->bigInteger('points');
             $table->timestamps();
         });
     }
